+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add stake, config and vote programs ([#57](https://github.com/LiteSVM/litesvm/pull/57)).
+- Implement blockhash and durable nonce checks ([#61](https://github.com/LiteSVM/litesvm/pull/61)).
 
 ### Changed